move pausable to its own file
